[mobile] Remove obsolete private migration review
