+++ lib/src/features/migration/screens/mobile/mobile_ironwood_migration_flow_screen.dart
@@ -30,14 +30,12 @@ import '../../models/mobile_ironwood_migration_status_entry.dart';
 import '../../providers/ironwood_migration_announcement_provider.dart';
 import '../../providers/ironwood_migration_coordinator_provider.dart';
 import '../../services/ironwood_migration_service.dart';
-import '../../widgets/ironwood_migration_shimmer_text.dart';
 import '../../widgets/mobile/mobile_ironwood_migration_attention.dart';
 import '../ironwood_migration_flow_screen.dart';
 
 part 'mobile_ironwood_migration_models.dart';
 part 'mobile_ironwood_migration_routes.dart';
 part 'mobile_ironwood_migration_intro_options.dart';
-part 'mobile_ironwood_migration_analyzing.dart';
 part 'mobile_ironwood_migration_review.dart';
 part 'mobile_ironwood_migration_live_states.dart';
 part 'mobile_ironwood_migration_preview_surfaces.dart';
@@ -48,9 +46,9 @@ part 'mobile_ironwood_migration_status_presentation.dart';
 part 'mobile_ironwood_migration_status_waiting.dart';
 part 'mobile_ironwood_migration_status_active.dart';
 part 'mobile_ironwood_migration_status_footer.dart';
+part 'mobile_ironwood_migration_status_metrics.dart';
 part 'mobile_ironwood_migration_step_scaffold.dart';
 part 'mobile_ironwood_migration_step_hero_process.dart';
 part 'mobile_ironwood_migration_step_options.dart';
-part 'mobile_ironwood_migration_step_plan.dart';
 part 'mobile_ironwood_migration_step_progress_parts.dart';
 part 'mobile_ironwood_migration_step_review_card.dart';
